requirements.txt
astroid==3.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dill==0.3.8
google-api-python-client==2.123.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
idna==3.6
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.0
platformdirs==4.2.0
playwright==1.42.0
pluggy==1.4.0
pyee==11.0.1
pylint==3.1.0
pytest==8.1.1
pytest-base-url==2.1.0
pytest-playwright==0.4.4
python-slugify==8.0.4
requests==2.31.0
text-unidecode==1.3
tomlkit==0.12.4
typing_extensions==4.10.0
urllib3==2.2.1
